--- modules/Presentation/README.md
@@ -1,10 +1,0 @@
-The only important file is presentation.py
-
-Helpers.py contains a few helper functions.
-
-test_data_generator.py simply returns test_data.
-
-Purgatory.py is for old / no longer relevant code / earlier versions of stuff.
-Its just to show the progress, and that I've been through a few iterations before I landed on whats in presentation.py
-Just so it doesnt look like I've spent a week on a few filter methods and a few simple .plot functions.
-For some reason I had a hard time getting the structure right.
